Refactoring

+++ CourseApp/AssaultRifle.cs
@@ -4,17 +4,17 @@
     {
         private ushort shoot = 0;
 
-        public AssaultRifle(string model, int year, string owner)
+        public AssaultRifle(string model, ushort year, string owner)
             : this(model, 0, year, owner, false)
         {
         }
 
-        public AssaultRifle(string model, int year, string owner, bool automaticShooting)
+        public AssaultRifle(string model, ushort year, string owner, bool automaticShooting)
             : this(model, 0, year, owner, automaticShooting)
         {
         }
 
-        public AssaultRifle(string model, float caliber, int year, string owner, bool automaticShooting)
+        public AssaultRifle(string model, float caliber, ushort year, string owner, bool automaticShooting)
         {
             this.Model = model;
             this.Owner = owner;
