+++ CourseApp/Pistol.cs
@@ -2,12 +2,12 @@
 {
     public class Pistol : Weapon
     {
-        public Pistol(string model, int year, string owner)
+        public Pistol(string model, ushort year, string owner)
             : this(model, 0, year, owner)
         {
         }
 
-        public Pistol(string model, float caliber, int year, string owner)
+        public Pistol(string model, float caliber, ushort year, string owner)
         {
             this.Model = model;
             this.Owner = owner;
